CrewAI_part/CrewAI6_여행스케줄Agent_update/requirements.txt
streamlit
crewai
langchain-openai
langchain
duckduckgo-search
langchain-anthropic
python-dotenv
langchain-community
